pyproject.toml
[project]
name = "realtime-assistant-poc"
version = "0.1.0"
description = "This project is a proof of concept using the Realtime API."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "chainlit==2.0.dev0",
    "openai==1.54.0",
    "websockets==13.1",
    "jinja2",
    "requests",
    "python-dotenv",
    "pydantic==2.10.1",
    "azure-storage-blob",
    "ipython",
    "ipykernel",
]
